# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trees:
    get:
      operationId: listTrees
      tags: [trees]
      summary: List trees (feed)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: string
        - name: province
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of trees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTrees"
    post:
      operationId: createTree
      tags: [trees]
      summary: Create a tree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreeBody"
      responses:
        "201":
          description: Created tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
  /trees/recent:
    get:
      operationId: getRecentTrees
      tags: [trees]
      summary: Get recent trees
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent trees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tree"
  /trees/{treeId}:
    get:
      operationId: getTree
      tags: [trees]
      summary: Get a tree
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
    patch:
      operationId: patchTree
      tags: [trees]
      summary: Patch a tree
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTreeBody"
      responses:
        "200":
          description: Updated tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
    delete:
      operationId: deleteTree
      tags: [trees]
      summary: Delete a tree
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trees/{treeId}/updates:
    get:
      operationId: getTreeUpdates
      tags: [trees]
      summary: Get tree updates
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tree updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreeUpdate"
    post:
      operationId: addTreeUpdate
      tags: [trees]
      summary: Add a tree update
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTreeUpdateBody"
      responses:
        "201":
          description: Created update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeUpdate"
  /trees/{treeId}/updates/{updateId}:
    delete:
      operationId: deleteTreeUpdate
      tags: [trees]
      summary: Delete a tree update
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
        - name: updateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trees/{treeId}/sun:
    post:
      operationId: sunTree
      tags: [trees]
      summary: Sun a tree
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sun result
          content:
            application/json:
              schema:
                type: object
                properties:
                  sunCount:
                    type: integer
                  userHasSunned:
                    type: boolean
                required:
                  - sunCount
                  - userHasSunned
    delete:
      operationId: unsunTree
      tags: [trees]
      summary: Unsun a tree
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsun result
          content:
            application/json:
              schema:
                type: object
                properties:
                  sunCount:
                    type: integer
                  userHasSunned:
                    type: boolean
                required:
                  - sunCount
                  - userHasSunned
...
components:
  schemas:
...
    PaginatedTrees:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: "#/components/schemas/Tree"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - trees
        - total
        - page
        - limit
    Tree:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        userPhotoUrl:
          type: string
          nullable: true
        photoUrl:
          type: string
        photoThumbnailUrl:
          type: string
          nullable: true
        plantName:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        species:
          type: string
          nullable: true
        plantedAt:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        locationName:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        mapsUrl:
          type: string
          nullable: true
        verificationBypassed:
          type: boolean
        photoStatus:
          type: string
        updateCount:
          type: integer
        sunCount:
          type: integer
        userHasSunned:
          type: boolean
        isWeeklyWinner:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - photoUrl
        - latitude
        - longitude
        - verificationBypassed
        - photoStatus
        - updateCount
        - sunCount
        - userHasSunned
        - isWeeklyWinner
        - createdAt
    TreeUpdate:
      type: object
      properties:
        id:
          type: integer
        treeId:
          type: integer
        photoUrl:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        photoStatus:
          type: string
        createdAt:
          type: string
      required:
        - id
        - treeId
        - photoStatus
        - createdAt
    CreateTreeBody:
      type: object
      properties:
        photoUrl:
          type: string
        photoThumbnailUrl:
          type: string
          nullable: true
        plantName:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        species:
          type: string
          nullable: true
        plantedAt:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        locationName:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        verificationBypassed:
          type: boolean
          nullable: true
        photoStatus:
          type: string
          nullable: true
      required:
        - photoUrl
        - latitude
        - longitude
    PatchTreeBody:
      type: object
      properties:
        plantName:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        species:
          type: string
          nullable: true
        plantedAt:
          type: string
          nullable: true
        locationName:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
    AddTreeUpdateBody:
      type: object
      properties:
        photoUrl:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: []
...